2_PYTHON_MLOPS/requirements.txt
numpy
pandas
scikit-learn
pyzmq
# MetaTrader5==3.0.32
# TA-Lib
hurst
pyyaml
joblib
streamlit
plotly